requirements.txt
streamlit
tensorflow-cpu
tensorflow
pillow
numpy
requests
tqdm
pexels-api
